Custom themes for the segmented control id not honor the label color
if the control is created with a custom theme, when the touch event is
processed, it checks for the current theme (if its ios7).

The problem was with the default and over colors - if the current theme
is not ios7, the control did not change the label colors. If it was
ios7, it always changed the over color to white