widget_neweditfield
 Added editfield_calibrationdb.lua that contains the database of
calibration data.
- Changed the textfield offsets for Android devices to account for the
adjusted height number, so now the offsets should be smaller
- Changed the fontScale alto for Corona emulator. Unfortunately the
Zooming of the emulator can not be retrieved, so to observe the real
fontScaling, you would need to Zoom In until the Zoom In is  greyed out
(actual pixel size). If the emulator is zoomed at any level, it will
display a wrong fontScale.
- Fixed the double onSubmit events
- Added androidKeyboardHeight to edit field defaults (defaults to 350
pixels) so user can change this setting in his app (like in main.lua
file)
- Added calibration string to send Email and save to file - this string
can be copy-pasted into editfield_calibrationdb.lua
- Fixed all known issues with scrolling of keyboard